requirements.txt
django-social-auth